refactor(charm): convert AAR interface to its own library with a custom event
This commit refactors the interactions with AAR interface into an early
charm library with the provider side of the interface implemented.
This commit also adds a custom event `client_registered` for AAR
interface which it emits after registering the client. This event can be
observed by the charm to take appropriate actions in response.